// packages/frontend/src/generated/types/AcceptanceCompositeResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AcceptanceItemResponse } from './AcceptanceItemResponse.ts'
import type { AcceptanceResponse } from './AcceptanceResponse.ts'

/**
 * @description Composite response DTO used by acceptance aggregate endpoints.
 */
export type AcceptanceCompositeResponse = AcceptanceResponse & {
  /**
   * @type array
   */
  items: AcceptanceItemResponse[]
}